A collection of classes for implementing web services in Objective-C
A collection of classes for making and/or parsing XMLRPC and SOAP requests.
This is intended to provide a framework for implementing web services,
so this release also contains initial support for parsing/editing
web service definitions (WSDL).

LICENSE: LGPL2 or later

WWW: http://www.gnustep.org/resources/downloads.php