This module builds on MooseX::Types::DateTime to add additional custom types and coercions. Since it builds on an existing type, all coercions and constraints are inherited.
The package name is left as is for legacy reasons: this module is
really a Type with coercions for DateTimeX::Easy. DateTimeX is just
a namespace for non-core or less-official DateTime modules.

WWW: http://search.cpan.org/dist/MooseX-Types-DateTimeX/

PR:		ports/135939 (5 of 6)
Submitted by:	Cezary Morga <cm at therek.net>