MuPDF is a lightweight PDF viewer and toolkit written in portable C.
The renderer in MuPDF is tailored for high quality anti-aliased graphics. It
renders text with metrics and spacing accurate to within fractions of a
pixel for the highest fidelity in reproducing the look of a printed page on
screen.

MuPDF has a small footprint. A binary that includes the standard Roman fonts
is only one megabyte. A build with full CJK support (including an Asian
font) is approximately five megabytes.

MuPDF has support for all non-interactive PDF 1.7 features, and the toolkit
provides a simple API for accessing the internal structures of the PDF
document. Example code for navigating interactive links and bookmarks,
encrypting PDF files, extracting fonts, images, and searchable text, and
rendering pages to image files is provided.

WWW: http://ccxvii.net/mupdf/

PR:		137559
Submitted by:	Martin Dieringer <martin.dieringer@gmx.de>