A Ruby interface to the PostgreSQL Relational Database Management System.
The 'pg' module is the newer module, that has been greatly improved, and
is almost a complete rewrite. It is not backwards compatible. Use this module
for newly written code. It should be more stable, less buggy, and has more
features.

LICENSE: BSD or GPL2

WWW:	http://rubyforge.org/projects/ruby-pg/