Add new port graphics/klatexformula:
KLatexFormula is a Qt4 program to easily get an image from a LaTeX
formula. Just enter a formula, click "Evaluate", then you can
drag&drop or copy the resulting image to another location (an
OpenOffice document, for example) or save it as an image (many
formats are available).

WWW:	http://klatexformula.sourceforge.net