. Add a port of log4net:
log4net is a tool to help the programmer output log statements to a variety of
output targets. In case of problems with an application, it is helpful to
enable logging so that the problem can be located. With log4net it is possible
to enable logging at runtime without modifying the application binary. The
log4net package is designed so that log statements can remain in shipped code
without incurring a high performance cost. It follows that the speed of logging
(or rather not logging) is crucial.

At the same time, log output can be so voluminous that it quickly becomes
overwhelming. One of the distinctive features of log4net is the notion of
hierarchical loggers. Using these loggers it is possible to selectively control
which log statements are output at arbitrary granularity.

log4net is designed with two distinct goals in mind: speed and flexibility.
Features:
* Support for multiple frameworks
* Output to multiple logging targets
* Hierarchical logging architecture
* XML Configuration
* Dynamic Configuration
* Logging Context
* Proven architecture
* Modular and extensible design
* High performance with flexibility

WWW: http://logging.apache.org/log4net/

PR:		141441
Submitted by:	Romain Tartiere <mono@FreeBSD.org>