Update client to v1.69 which is the final release from BioWare.
Remove ARCH requirement for i386; let the install of the Linux base
determine if the port is allowed or not.

Detect if the original or Diamond game files were installed in
games/nwndata to install the appropriate client.

Add an option to install the NWMovies/BinkPlayer patch to play in-game
movies for the Diamond client.  This includes a rewritten script (from
Perl to shell) to remove the need for Linux Perl to run it.  The script
includes a method to skip movies, especially the intro movies, as noted in
pkg-message.  Default to off.

In the nwn script, remove dead links in and rebuild ${HOME}/.nwn.  This
allows moving between the original and Diamond editions without confusing
(resulting in segmentation faults) the client.

Set SDL_AUDIODRIVER to dsp by default to remove warnings from SDL
concerning audio setup.

Disallow core files as these are commonly seen when the game exits.
Fortunately, the segmentation fault does not affect play nor the
configuration files.

Approved by:	wxs