Remove CONFLICTS for the old pine stuff, it's been long enough
Fix the problem mentioned in the PR related to a feature of the port
that is useful, or dangerous depending on how you look at it. :)
If you run 'alpine -conf > file' it will merge in values from your
existing global conf file (/usr/local/etc/alpine.conf) and spit out a
new file with any new features added. The port used this feature when
I took it over, and I maintained that behavior because it is useful.
However, it is different from the traditional behavior of installing
a clean foo.conf.sample file, and maintaining foo.conf only if it
differs from the sample.

My solution to this problem is different than the PR's, but does not
involve patching the source. Using the pkg-install file and taking
advantage of the default behavior of the alpine -conf feature I have
created the best of both worlds, a clean .sample, and merging in local
changes if they exist.

Since I'm changing stuff anyway, do both sides of the process in a
more security-conscious way.

Bump PORTREVISION since the package is now different

PR:		ports/148859
Submitted by:	Ganael Laplanche <ganael.laplanche@martymac.com>