New port math/octave-forge-fuzzy-logic-toolkit
This is fuzzy-logic-toolkit.

A mostly MATLAB-compatible fuzzy logic toolkit for Octave.

The octave-forge package is the result of The GNU Octave Repositry project,
which is intended to be a central location for custom scripts, functions and
extensions for GNU Octave. contains the source for all the functions plus
build and install scripts.

WWW: http://octave.sourceforge.net/

PR:		ports/157970
Submitted by:	stephen
Approved by:	maho (mentor)